requirements.txt
flask==2.0.1
line-bot-sdk==2.4.2
openai==1.3.5
python-dotenv==0.19.0
gunicorn==20.1.0
SpeechRecognition==3.10.0
pydub==0.25.1
httpx==0.24.1
Werkzeug==2.0.1
redis==5.0.1
Pillow==9.5.0